SS-577: initial changes to provide duty roster week view
