+++ web/src/components/DutyRoster/components/DutyRosterHeader.vue
@@ -7,7 +7,7 @@
 				</b-navbar-nav>
 				<b-navbar-nav class="custom-navbar">
                     <b-col>
-                        <b-row style="margin:.25rem auto .25rem auto; width:73%;">
+                        <b-row style="margin:.25rem auto .25rem auto; width:100%;">
                             <b-button style="height: 2rem;" size="sm" variant="secondary" @click="previousDateRange" class="my-0"><b-icon-chevron-left /></b-button>
                             <b-form-datepicker
                                     class="my-0 py-0 mx-1"
@@ -25,12 +25,25 @@
                             <b-button style="height: 2rem;" size="sm" variant="secondary" @click="nextDateRange" class="my-0"><b-icon-chevron-right/></b-button>
                         </b-row>
                         <b-row style="margin:0 0 .25rem auto;">
-                            <div class="bg-white" style=" border-radius:5px; width:10rem;">{{selectedDate|beautify-date-weekday}}</div>
-                        </b-row>
+                            <div v-if="activetab=='Day'" class="bg-white" style="transform:translate(-1.3rem,0); border-radius:3px; width:10rem;">{{selectedDate|beautify-date-weekday}}</div>
+							<div v-else class="bg-white" style="transform:translate(-5rem,0); border-radius:3px; width:17rem;">{{selectedDateBegin|beautify-date-weekday}} - {{selectedDateEnd|beautify-date-weekday}}</div>
+						</b-row>
                     </b-col>
                 </b-navbar-nav>
 				<b-navbar-nav>
-					<div style="width:11rem;" class="text-white ml-2 mr-auto"></div>
+					<b-tabs nav-wrapper-class = "bg-primary text-dark"
+							active-nav-item-class="text-uppercase font-weight-bold text-warning bg-primary"                     
+							pills
+							no-body
+							class="mx-3">
+						<b-tab 
+							v-for="(tabMapping, index) in tabs" 
+							:key="index"                 
+							:title="tabMapping"                 
+							v-on:click="tabChanged(tabMapping)" 
+							v-bind:class="[ activetab === tabMapping ? 'active mb-0' : 'mb-0' ]"
+							/>
+					</b-tabs>
 				</b-navbar-nav>
 			</b-navbar>
 		</header>
@@ -277,8 +290,13 @@
 		
 		@Prop({required: true})
 		runMethod!: any
+
+		activetab = 'Day';
+		tabs =['Day', 'Week']
 		
 		selectedDate = '';
+		selectedDateBegin = '';
+		selectedDateEnd = '';
 		datePickerOpened = false;
 		userIsAdmin = false;
 
@@ -337,7 +355,7 @@
 			this.selectedDate = moment().format().substring(0,10);			
 			this.loadNewDateRange();
 		}
-
+		
 		public weekdaysChanged(){
 			Vue.nextTick(()=>{
 				this.allDaysSelected = this.selectedDays.length==7? true: false
@@ -588,7 +606,6 @@
 			}
 		}
 
-
 		public createAssignment() {
 
 			this.assignment.sunday = this.selectedDays.includes(0)
@@ -629,27 +646,37 @@
 				})
 		}
 
+		public tabChanged(tabInfo){
+			this.activetab = tabInfo;
+			this.loadNewDateRange();
+		}
+
         
         public dateChanged(event) {
 			if(this.datePickerOpened)this.loadNewDateRange();
 		}
 
-		public nextDateRange() {			
-			this.selectedDate = moment(this.selectedDate).add(1, 'days').format().substring(0,10);
+		public nextDateRange() {
+			const days =(this.activetab == 'Day')? 1 :7;		
+			this.selectedDate = moment(this.selectedDate).add(days, 'days').format().substring(0,10);
 			this.loadNewDateRange(); 
 		}
 
 		public previousDateRange() {
-			this.selectedDate = moment(this.selectedDate).subtract(1, 'days').format().substring(0,10);
+			const days =(this.activetab == 'Day')? 1 :7;
+			this.selectedDate = moment(this.selectedDate).subtract(days, 'days').format().substring(0,10);
 			this.loadNewDateRange();
 		}
 
 		public loadNewDateRange() {
-			const beginTime = moment(this.selectedDate).startOf('day').format()
-			const endTime = moment(this.selectedDate).endOf('day').format();
-			const dateRange = {startDate: beginTime, endDate: endTime}
+
+			const dateType = (this.activetab == 'Day')?'day':'week'
+			this.selectedDateBegin = moment(this.selectedDate).startOf(dateType).format()
+			this.selectedDateEnd = moment(this.selectedDate).endOf(dateType).format();
+			const dateRange = {startDate: this.selectedDateBegin, endDate: this.selectedDateEnd}
+			console.log(dateRange)
 			this.UpdateDutyRangeInfo(dateRange);
-			this.$emit('change'); 
+			this.$emit('change',this.activetab); 
 		}
 
 		public timeFormat(value , event) {
